Remove FunnelStep and move action step into separate file
