+++ posthog/models/__init__.py
@@ -1,10 +1,11 @@
-from .user import User, UserManager
-from .event import Event
-from .element_group import ElementGroup
-from .action import Action, ActionStep
+from .action import Action
+from .action_step import ActionStep
 from .dashboard import DashboardItem
-from .team import Team
-from .element import Element
-from .person import Person, PersonDistinctId
-from .funnel import Funnel
 from .cohort import Cohort
+from .element import Element
+from .element_group import ElementGroup
+from .event import Event
+from .funnel import Funnel
+from .person import Person, PersonDistinctId
+from .team import Team
+from .user import User, UserManager
